% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAREG.R
\name{STAREG}
\alias{STAREG}
\title{Replicability analysis of two multiple testing studies providing FDR control}
\usage{
STAREG(
  pvals1,
  pvals2,
  minIter = 3,
  maxIter = 100,
  est.pi0 = FALSE,
  pi0.const = 0.9,
  tol = 0.001,
  trace = FALSE
)
}
\arguments{
\item{pvals1}{a numeric vector of p-values from study 1.}

\item{pvals2}{a numeric vector of p-values from study 2.}

\item{minIter}{an integer value indicating the minimum number of iterations.}

\item{maxIter}{an integer value indicating the maximum number of iterations.}

\item{est.pi0}{a logical value indicating whether to estimate the proportion of true nulls in each study for the initialization of xi's or just specify the value as 0.9. Default is FALSE.}

\item{pi0.const}{a numeric number between 0 and 1 to initialize xi's if est.pi0 is false.}

\item{tol}{a numeric value giving the tolerance in the relative change in the log likelihood below which the EM algorithm is considered to be converged. Default is set to 1e-3.}

\item{trace}{a logical value indicating whether to print out the EM process.}
}
\value{
a list with the following elements:
\item{call}{the function call made.}
\item{lfdr.rep}{a numeric vector of the estimated values of joint local FDR.}
\item{fdr.rep}{a numeric vector of the adjusted p-values for FDR control based on the joint local FDR.}
\item{xi00}{a numeric value of the estimated probability for the joint hidden state of (0, 0)}
\item{xi01}{a numeric value of the estimated probability for the joint hidden state of (0, 1)}
\item{xi10}{a numeric value of the estimated probability for the joint hidden state of (1, 0)}
\item{xi11}{a numeric value of the estimated probability for the joint hidden state of (1, 1)}
\item{f1}{a vector of the estimated densities for the p-value under the alternatives in study 1}
\item{f2}{a vector of the estimated densities for the p-value under the alternatives in study 1}
\item{loglik}{a numeric vector of the log likelihood at each step}
\item{converge}{a list containing the convergence information. \code{code}: 1 - converged, 0 - not converged; \code{iter}: the number of iterations performed.}
}
\description{
Replicability analysis of two multiple testing studies providing FDR control
}
